Code Cleanup (#824)
* refactor: clean up code

- "merges" slash and interaction contexts into super class
- fixes a bug in timestamp logic of ccv3

* refactor: clean up code pt2

* refactor: super class for new commands & missing override anno

* refactor: weed out unused stuff or suppress warning is used/potentially important
refactor: deal with or suppress probable bugs
refactor(command permission): use predicate
fix(slash/context command): use user permission has default (as was before)
add(cotext command) modal check
refactor(commands): remove guild only meta tag

* refactor: post commit checks & remove unused string

* refactor(all): prefer member suppression -> class suppression
fix(food item): castable required methods got nuked by merge
fix(inventory): fix serialization

* fix(suppression): apply member -> class to some classes missed

* refactor(suppression): apply review

* refactor(managed mongo obj): save -> insertOrReplace (more accurate)

---------

+++ src/main/java/net/kodehawa/mantarobot/commands/CurrencyActionCmds.java
@@ -323,7 +323,7 @@ public class CurrencyActionCmds {
             var amount = 1 + random.nextInt(item.getDiamondIncrease());
             if (!player.fitsItemAmount(ItemReference.DIAMOND, amount)) {
                 message += "\n" + languageContext.get("commands.mine.diamond.overflow").formatted(amount);
-                overflowMoney += Math.round((ItemReference.DIAMOND.getValue() * 0.9) * amount);
+                overflowMoney += (int) Math.round((ItemReference.DIAMOND.getValue() * 0.9) * amount);
             } else {
                 player.processItem(ItemReference.DIAMOND, amount);
                 message += "\n" + EmoteReference.DIAMOND + languageContext.get("commands.mine.diamond.success").formatted(amount);
@@ -392,7 +392,7 @@ public class CurrencyActionCmds {
             if (!player.fitsItemAmount(itemGem, selectedGem.getAmount())) {
                 message += "\n" + languageContext.get("commands.mine.gem.overflow")
                         .formatted(itemGem.getEmojiDisplay() + " x" + selectedGem.getAmount());
-                overflowMoney += Math.round((itemGem.getValue() * 0.9) * selectedGem.getAmount());
+                overflowMoney += (int) Math.round((itemGem.getValue() * 0.9) * selectedGem.getAmount());
             } else {
                 player.processItem(selectedGem);
 
@@ -432,7 +432,7 @@ public class CurrencyActionCmds {
 
             if (!player.canFitItem(gem)) {
                 message += "\n" + languageContext.get("commands.mine.sparkle.overflow");
-                overflowMoney += Math.round(gem.getValue() * 0.9);
+                overflowMoney += (int) Math.round(gem.getValue() * 0.9);
             } else {
                 player.processItem(gem, 1);
                 message += "\n" + EmoteReference.MEGA +
